Reworked deadzone functionality and fixed scrolling
There are now deadzones for moving and tapping, where tapping is a lot larger, and moving is smaller.
deadZoneMove is for countering jitter in the initial finger movement, while deadZoneTap tries to minimize unintended double-tapping when moving the cursor at higher speeds (tapping the finger fast back and forth on screen).